+++ geopackage-sdk/src/main/java/mil/nga/geopackage/factory/GeoPackageManagerImpl.java
@@ -781,22 +781,34 @@ class GeoPackageManagerImpl implements GeoPackageManager {
      */
     @Override
     public GeoPackage open(String database) {
+        return open(database, true);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public GeoPackage open(String database, boolean writable) {
 
         GeoPackage db = null;
 
         if (exists(database)) {
             GeoPackageCursorFactory cursorFactory = new GeoPackageCursorFactory();
             String path = null;
-            boolean writable = true;
-            SQLiteDatabase sqlite;
+            SQLiteDatabase sqlite = null;
             GeoPackageMetadata metadata = getGeoPackageMetadata(database);
             if (metadata != null && metadata.isExternal()) {
                 path = metadata.getExternalPath();
-                try {
-                    sqlite = SQLiteDatabase.openDatabase(path,
-                            cursorFactory, SQLiteDatabase.OPEN_READWRITE
-                                    | SQLiteDatabase.NO_LOCALIZED_COLLATORS);
-                } catch (Exception e) {
+                if(writable){
+                    try {
+                        sqlite = SQLiteDatabase.openDatabase(path,
+                                cursorFactory, SQLiteDatabase.OPEN_READWRITE
+                                        | SQLiteDatabase.NO_LOCALIZED_COLLATORS);
+                    } catch (Exception e) {
+                        Log.e(GeoPackageManagerImpl.class.getSimpleName(), "Failed to open database as writable: " + database, e);
+                    }
+                }
+                if(sqlite == null){
                     sqlite = SQLiteDatabase.openDatabase(path,
                             cursorFactory, SQLiteDatabase.OPEN_READONLY
                                     | SQLiteDatabase.NO_LOCALIZED_COLLATORS);